fix(backend): reestructuracion y documentacion de las entidades
